feat: change vector type from int to short

+++ src/global/Vector100Dtype.java
@@ -2,22 +2,22 @@ package global;
 
 public class Vector100Dtype {
 
-  private final static int MIN_VAL = -10000;
-  private final static int MAX_VAL = 10000;
-  private int[] dimension;
+  private final static short MIN_VAL = -10000;
+  private final static short MAX_VAL = 10000;
+  private short[] dimension;
 
   public Vector100Dtype() {
-    dimension = new int[100];
+    dimension = new short[100];
   }
 
-  public Vector100Dtype(int[] dimension) {
-    this.dimension = new int[100];
+  public Vector100Dtype(short[] dimension) {
+    this.dimension = new short[100];
     setDimension(dimension);
   }
-  public int[] getDimension() {
+  public short[] getDimension() {
     return dimension;
   }
-  public void setDimension(int[] dimension) {
+  public void setDimension(short[] dimension) {
     for (int i = 0; i < 100; i++) {
       if (dimension[i] > MAX_VAL)
         this.dimension[i] = MAX_VAL;
